+++ src/aui-button/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-1873](https://issues.liferay.com/browse/AUI-1873) aui button visible does not output proper css class
 
 ## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)